fix(main/libluajit): Fix mismatch between version and what is built
Background: luajit uses rolling releases - see
https://luajit.org/download.html

The recommended approach is to make frequent snapshots, which the termux
package is currently taking from the 2.1 branch.

The version scheme follows what arch linux is doing:
https://archlinux.org/packages/extra/x86_64/luajit/

That means that the version number "2.1.1716656478" refers to the "2.1"
branch, with "1716656478" being the timestamp of the commit built (as
displayed by `git show -s --format=%ct`).

The problem is that the termux package was hardcoded to build a specific
commit from 2021-03-10:
https://github.com/LuaJIT/LuaJIT/commit/787736990ac3b7d5ceaba2697c7d0f58f77bb782

We now instead look up the commit from the timestamp in the version string,
so that what is built is correct, and auto updates will start working as expected.