fix(scripts/termux_step_install_license): fix logic error
Move counter advance to the end of the while loop.
This fixes a logic error where the counter wouldn't advance
if the first license was "generic", and the second one was
author specific. This would cause `cp` to attempt to copy
through the dangling symlink to the generic license
in `termux-licenses` which fails and dies.