fix(main/nushell): Build with default features
Clipboard integration historically hasn't worked on Termux anyway,
so cargo's features were chosen to explicitly opt into a feature that
included all default features except clipboard support.

9f0a6bee57dd96fd608ffefb166686111d4682d8 upgraded to 0.98.0,
which removed support for clipboard integration, and thus the need
for that feature flag went away. Unfortunately, it removed the
`default-no-clipboard` feature, but not the `no-default-features` flag,
so the resulting binary is missing a lot of stuff, like plug-in support,
sqlite, and others.

This fixes that by removing the `--no-default-features` flag.